rndobj/Mesh: CompressedVertex_Xbox field types float/uint -> SaveCompressedVertex +1 @100%
CompressedVertex_Xbox.mPosX/Y/Z were typed int; retail (and DC3) type them
float, and mNormal/mTangent/mBinormal/mBoneIndices/mBoneWeights as unsigned int.
The wrong types made BinStream::operator<< resolve to the int writer (lwz/stw
codegen) instead of the float writer (lfs/stfs). Fixing the types + chaining the
first three inserts (bs << x << y << z, per DC3) flips SaveCompressedVertex to
100%. FillCompressedVertex casts already type-compatible; DxMesh (other includer)
unaffected. Zero regressions (fresh whole-binary report).

+++ src/system/rndobj/MeshVertCompress.h
@@ -4,15 +4,15 @@
 #include "utl/BinStream.h"
 
 struct CompressedVertex_Xbox {
-    int mPosX;
-    int mPosY;
-    int mPosZ;
+    float mPosX;
+    float mPosY;
+    float mPosZ;
     int mColor; // 0xc - packed color
-    int mNormal;
-    int mTangent;
-    int mBinormal;
-    int mBoneIndices;
-    int mBoneWeights;
+    unsigned int mNormal;
+    unsigned int mTangent;
+    unsigned int mBinormal;
+    unsigned int mBoneIndices;
+    unsigned int mBoneWeights;
 };
 
 void PackVector(
